feat(ntfy): l'auto-heberge sait notifier, sans passer par Google
FCM lie la notification au BINAIRE de l'app : elle n'atteint cet APK que si
elle part d'un compte de service du projet Firebase avec lequel il a ete
compile — celui de l'editeur. Un serveur chez soi n'avait donc que de
mauvaises portes : sa propre cle n'atteint pas cet APK, celle de l'editeur
ne peut pas voyager dans un depot public, et relayer par le nuage
contredirait la raison meme de s'auto-heberger.

ntfy n'a aucun de ces problemes. Du HTTP, aucun compte, l'utilisateur
installe l'app ntfy — et ca marche aussi avec une instance qu'on heberge
soi-meme. C'est la convention du monde auto-heberge.

## L'ACTION ne change pas

Une regle dit « previens-moi ». COMMENT est une capacite du serveur, pas un
detail de la regle. Le nuage la sert par FCM, le jumeau par ntfy, et le
langage des regles reste le meme des deux cotes — c'est ce qui garde les
trente-quatre fixtures partagees valables ici.

L'expediteur s'enregistre donc sous `TYPE_PUSH`, et la source unique fait le
reste : `allowedActionTypes = actionSenders.keys` ouvre la creation des
regles push toute seule, sans qu'une seconde liste ait a etre tenue a jour.
Verifie en vrai — la regle a ete acceptee, alors qu'elle etait refusee une
heure plus tot.

## Deux questions qui se ressemblaient, et que ntfy separe

« Sais-tu livrer une action PUSH ? » — oui desormais.

« Veux-tu mon jeton d'appareil ? » — non, et c'est CETTE question que l'app
pose a `/api/auth/info`. Elle ne l'appelle que pour decider d'envoyer son
jeton FCM a `/api/push-tokens`, une route que ce serveur n'a pas et n'en
veut pas : ntfy ne passe pas par l'app, il passe par l'app ntfy. Repondre
`true` ferait envoyer un jeton inutilisable a chaque connexion, et
demanderait la permission d'afficher des notifications qui n'arriveraient
jamais par ce chemin.

FCM confondait les deux tant qu'il etait le seul canal.

## Le sujet EST le secret

Sur `ntfy.sh` public, quiconque connait le nom du sujet recoit ses
notifications ET peut en publier. C'est le modele de ntfy, pas un defaut,
mais il se dit : le panneau refuse une barre ou un espace, et la
documentation rappelle qu'un sujet doit etre long et imprevisible. Une
instance privee avec jeton existe pour qui veut mieux ; le jeton n'est
jamais RENDU par l'API, seulement sa presence.

## Un defaut d'affichage, trouve au passage

La premiere notification disait « Il fait 37.400001525878906 C ». Une trame
numerique porte quatre octets ; elargis en `Double`, ils portent l'ecart de
la representation. Tu l'avais deja vu en production le 9 septembre :
« Current value: 37.379913330078125 ».

`{{value}}` rend desormais un float32 : la plus courte chaine qui revient
exactement au meme float32, donc « 37.4 ». Ce n'est pas un arrondi
d'affichage, c'est la valeur telle que la carte l'a envoyee, sans les
chiffres que l'elargissement a inventes.

LE NUAGE A LE MEME DEFAUT, a l'identique — `render` est le meme fichier des
deux cotes. Le correctif y est a porter.

## Eprouve

Onze epreuves sur l'expediteur, trois sur le gabarit, et la chaine complete
sur un serveur reel : la carte envoie 37,4, la regle tire, et la
notification arrive sur ntfy.sh, relue depuis le sujet. 610 tests.

+++ src/main/kotlin/com/jeanloickdt/automation/v2/AutomationEngine.kt
@@ -660,9 +660,21 @@ class AutomationEngine(
             is TypedValue.Int -> value.value.toString()
             // Un entier reste un entier a l'affichage : « 31.0 °C » se lit
             // moins bien que « 31 », et le fil rend tout numerique en double.
+            // Rendu comme un FLOAT32, parce que c'est ce que le fil porte.
+            //
+            // Une trame numerique transporte quatre octets. Elargis en
+            // `Double`, ils portent l'ecart de la representation :
+            // `37.4f` devient `37.400001525878906`, et c'est ce que
+            // l'utilisateur lisait dans sa notification — vu en production le
+            // 9 septembre 2026, « Current value: 37.379913330078125 ».
+            //
+            // `toFloat().toString()` rend la plus COURTE chaine qui revient
+            // exactement au meme float32, donc « 37.4 ». Ce n'est pas un
+            // arrondi d'affichage : c'est la valeur telle que la carte l'a
+            // envoyee, sans les chiffres que l'elargissement a inventes.
             is TypedValue.Float ->
                 if (value.value % 1.0 == 0.0) value.value.toLong().toString()
-                else value.value.toString()
+                else value.value.toFloat().toString()
             is TypedValue.Text -> value.value
         }
         return text.replace(TEMPLATE_VALUE, rendered)
